backend/requirements.txt
fastapi>=0.104.1,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
sqlalchemy>=2.0.23,<3.0.0
alembic>=1.12.1,<2.0.0
psycopg2-binary>=2.9.9,<3.0.0
python-multipart>=0.0.6,<1.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt==4.0.1  # pin to avoid passlib incompatibility with newer bcrypt
python-dotenv>=1.0.0,<2.0.0
redis>=5.0.1,<6.0.0
celery>=5.3.4,<6.0.0
crewai>=0.28.0,<0.30.0  # Using a more recent stable version
openai>=1.3.7,<2.0.0
weasyprint>=60.1,<61.0.0
stripe>=7.6.0,<8.0.0
pypdf2>=3.0.1,<4.0.0
python-docx>=1.1.0,<2.0.0
beautifulsoup4>=4.12.2,<5.0.0
boto3>=1.29.7,<2.0.0
